CTA-06-02: произвел небольшой рефакторинг - создал отдельную SelectorViewModel.kt, исправил баг с некорректным названием начального списка.
